Profiler 19.03.26 (#74)
* refactor(shell): improve get_source_dir() readability

* refactor(shell): move Project parameter to constructor

* fix(cli): add setup environment for profiler

* feat(profiler): make comprehensive UI messages

* fix(shell): fix missing newlines in paramiko

* fix(profiler): avoid copying archive if not exist

* fix(cli): fix incorrect profile_all cwd

* refactor(profiler): update profile_all
Now returns a list of executables for which all profiler steps were completed successfuly.

* feat(profiler): add cleaning-up profiler's files

* refactor(profiler): add events parameter to perf_record_collect()

* fix(shell): fix false error logging

* fix(shell): add handling missing known_hosts file

* fix(shell): make local copying not use sshpass

* fix(profiler): remove redundant pwd command run

* feat(ci): add CI artifacts uploading support

* fix(tests): use uv

* fix(profiler): print available events on unknown event

* fix(shell): local machine get_home() didn't return evaluated value

* fix(ci): fix fail on unit tests doesn't return error

+++ tests/integration/conftest.py
@@ -26,9 +26,10 @@ def clone_repo():
 
     yield _clone_repo
 
-    for path in repo_paths:
-        if os.path.exists(path):
-            shutil.rmtree(path, ignore_errors=True)
+    if not os.getenv("CI"):
+        for path in repo_paths:
+            if os.path.exists(path):
+                shutil.rmtree(path, ignore_errors=True)
 
 
 @pytest.fixture
@@ -45,9 +46,10 @@ def create_working_space():
 
     yield _create_temp_file
 
-    for dir_ in working_dirs:
-        if os.path.exists(dir_):
-            shutil.rmtree(dir_, ignore_errors=True)
+    if not os.getenv("CI"):
+        for dir_ in working_dirs:
+            if os.path.exists(dir_):
+                shutil.rmtree(dir_, ignore_errors=True)
 
 
 @pytest.fixture
@@ -62,4 +64,5 @@ def _docker_compose():
 
     yield compose
 
-    compose.stop()
+    if not os.getenv("CI"):
+        compose.stop()
